+++ Sora/MediaChannel.swift
@@ -108,8 +108,8 @@ public final class MediaChannel {
     }
 
     /// メディアチャンネルの内部で利用している RTCPeerConnection
-    public var native: RTCPeerConnection {
-        peerChannel.context.nativeChannel
+    public var native: RTCPeerConnection? {
+        peerChannel.nativeChannel
     }
 
     /**